Openfin epics refactored into openfin folder
