+++ src/client/src/configureStore.ts
@@ -5,11 +5,11 @@ import { combineEpics, createEpicMiddleware } from 'redux-observable'
 import { ApplicationDependencies } from './applicationServices'
 import rootReducer, { GlobalState } from './combineReducers'
 import { linkEpic } from './linkEpic'
-import { openfinEpic } from './openfinEpics'
 import { compositeStatusServiceEpic } from './operations/compositeStatusService'
 import { connectionStatusEpic } from './operations/connectionStatus'
 import { pricingServiceEpic } from './operations/pricing'
 import { referenceServiceEpic } from './operations/referenceData'
+import { openfinEpic } from './services/openFin/epics'
 import { analyticsServiceEpic } from './ui/analytics'
 import { blotterEpic } from './ui/blotter/'
 import { popoutEpic } from './ui/common/popout/popoutEpic'
