+++ src/client/src/services/openFin/epics.ts
@@ -1,12 +1,16 @@
 import { combineEpics } from 'redux-observable'
 import { connectAnalyticsServiceToOpenFinEpic } from './epics/analyticsServiceEpic'
 import { connectBlotterServiceToOpenFinEpic } from './epics/blotterServiceEpic'
+import { closePositionEpic } from './epics/closePosition'
 import { connectCurrencyChartToOpenFinEpic } from './epics/currencyChartEpic'
+import { publishPriceToOpenFinEpic } from './epics/publishPrice'
 import { connectTradeExecutedToOpenFinEpic } from './epics/tradeExecutedEpic'
 
-export const openfinServiceEpics = combineEpics(
+export const openfinEpic = combineEpics(
   connectAnalyticsServiceToOpenFinEpic,
   connectBlotterServiceToOpenFinEpic,
   connectCurrencyChartToOpenFinEpic,
-  connectTradeExecutedToOpenFinEpic
+  connectTradeExecutedToOpenFinEpic,
+  publishPriceToOpenFinEpic,
+  closePositionEpic
 )
